+++ azure/eventhub/common.py
@@ -5,6 +5,7 @@
 
 import datetime
 import time
+import json
 
 from uamqp import Message, BatchMessage
 from uamqp import types, constants, errors
@@ -87,7 +88,6 @@ class EventData(object):
                 raise ValueError("EventData cannot be None.")
             else:
                 self.message = Message(body, properties=self.msg_properties)
-
 
     @property
     def sequence_number(self):
@@ -188,7 +188,45 @@ class EventData(object):
 
         :rtype: bytes or Generator[bytes]
         """
-        return self.message.get_data()
+        try:
+            return self.message.get_data()
+        except TypeError:
+            raise ValueError("Message data empty.")
+
+    def body_as_str(self, encoding='UTF-8'):
+        """
+        The body of the event data as a string if the data is of a
+        compatible type.
+
+        :param encoding: The encoding to use for decoding message data.
+         Default is 'UTF-8'
+        :rtype: str
+        """
+        data = self.body
+        try:
+            return "".join(b.decode(encoding) for b in data)
+        except TypeError:
+            return str(data)
+        except:
+            pass
+        try:
+            return data.decode(encoding)
+        except Exception as e:
+            raise TypeError("Message data is not compatible with string type: {}".format(e))
+
+    def body_as_json(self, encoding='UTF-8'):
+        """
+        The body of the event loaded as a JSON object is the data is compatible.
+
+        :param encoding: The encoding to use for decoding message data.
+         Default is 'UTF-8'
+        :rtype: dict
+        """
+        data_str = self.body_as_str()
+        try:
+            return json.loads(data_str)
+        except Exception as e:
+            raise TypeError("Event data is not compatible with JSON type: {}".format(e))
 
 
 class Offset(object):