+++ HISTORY.rst
@@ -15,6 +15,7 @@ Release History
   - The `lease_container_name` argument now defaults to `"eph-leases"` if not specified.
 
 - Fix for clients failing to start if run called multipled times (issue #64).
+- Added convenience methods `body_as_str` and `body_as_json` to EventData object for easier processing of message data.
 
 
 1.0.0 (2018-08-22)